feat(android): implement immediate SMS sync for premium users
- Extract SMS sync logic from `SmsSyncWorker` to `SmsCloudSynchronizer`.
- Inject `SmsCloudSynchronizer` into `SmsSyncWorker` and `SmsSyncManager`.
- Update `SmsSyncManager` to trigger immediate sync via coroutines for premium users when messages change or settings update.
- Maintain WorkManager-based background sync for free users.
- Ensure premium users get real-time SMS updates on the web app.
